[SPARK-26029][BUILD][2.4] Bump previousSparkVersion in MimaBuild.scala to be 2.3.0
## What changes were proposed in this pull request?

Although it's a little late, we should still update mima for branch 2.4, to avoid future breaking changes.

Note that, when merging, we should forward port it to master branch, so that the excluding rules are still in `v24excludes`.

TODO: update the release process document to mention about mima update.

## How was this patch tested?

N/A

Closes #23015 from cloud-fan/mima-2.4.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>